requirements.txt
asgiref==3.3.4
Django==3.2
django-filter==2.4.0
djangorestframework==3.12.4
psycopg2==2.8.4
python-decouple==3.4
pytz==2021.1
rispy==0.6.0
sqlparse==0.4.1
